don't execute code unnecessarily if nothing changes

+++ app/src/main/java/de/westnordost/streetcomplete/data/osm/edits/ElementEditsController.kt
@@ -70,6 +70,7 @@ import javax.inject.Singleton
      *  the undo history */
     fun deleteSyncedOlderThan(timestamp: Long): Int {
         val edits = editsDB.getSyncedOlderThan(timestamp)
+        if (edits.isEmpty()) return 0
         val result = editsDB.deleteAll(edits.map { it.id })
         onDeletedEdits(edits)
         return result
